Moved functions from server.js to Abtin
;(
