+++ database/database.js
@@ -6,12 +6,6 @@
 */
 
 import mysql from 'mysql2';
-import dotenv from 'dotenv';
-
-// Load environment variables from .env file
-dotenv.config({ path: './hemmeligheder.env' });
-
-console.log('Loaded Environment Variables:', process.env);
 
 /* Connection pooling to improve performance, avoid overloading MYSQL server with too many connections
   creates a pool of reusable connections to the database - instead of creating a new connection for each query*/
@@ -23,25 +17,5 @@ const dbPool = mysql.createPool({
     password: process.env.MYSQL_PASSWORD,
     database: process.env.MYSQL_DATABASE,
 }).promise();    //promise() allows use promise API version of mysql
-
-dbPool.getConnection((err, connection) => {
-    if (err) {
-        console.error('Error connecting to the database:', err);
-    } else {
-        console.log('Database connected successfully!');
-        connection.release();
-    }
-});
-
-const result = await dbPool.query("SELECT * FROM customer")
-
-console.log(result);
-
-console.log('Environment Variables:', {
-    host: process.env.MYSQL_HOST,
-    user: process.env.MYSQL_USER,
-    password: process.env.MYSQL_PASSWORD,
-    database: process.env.MYSQL_DATABASE
-});
 
 export default dbPool; // Export the connection pool for use in other files
