+++ routes/account-routes.js
@@ -1,0 +1,9 @@
+import { createAccount, memberLogin, useTempPassword } from "../controllers/account-controller";
+
+
+// Route to handle account creation
+app.post('/create-account', createAccount);
+
+app.post('/login', memberLogin);
+
+app.post("/send-temp-password", useTempPassword);
